requirements.txt
discord.py
youtube-search-python
git+https://github.com/nficano/pytube.git
youtube_dl
asyncio
psutil
gpiozero